Move stackdriver stuff into its own file
