+++ cmd/ipfs/main.go
@@ -9,13 +9,10 @@ import (
 	"os"
 	"path/filepath"
 	"runtime/pprof"
-	"strconv"
 	"strings"
 	"time"
 
-	"contrib.go.opencensus.io/exporter/stackdriver"
-	"go.opencensus.io/trace"
-
+	cmds "github.com/ipfs/go-ipfs-cmds"
 	util "github.com/ipfs/go-ipfs/cmd/ipfs/util"
 	oldcmds "github.com/ipfs/go-ipfs/commands"
 	core "github.com/ipfs/go-ipfs/core"
@@ -26,7 +23,7 @@ import (
 	fsrepo "github.com/ipfs/go-ipfs/repo/fsrepo"
 
 	osh "github.com/Kubuxu/go-os-helper"
-	"github.com/ipfs/go-ipfs-cmds"
+	cmds "github.com/ipfs/go-ipfs-cmds"
 	"github.com/ipfs/go-ipfs-cmds/cli"
 	"github.com/ipfs/go-ipfs-cmds/http"
 	config "github.com/ipfs/go-ipfs-config"
@@ -89,42 +86,6 @@ func main() {
 		exitCode = mainRet()
 	})
 	os.Exit(exitCode)
-}
-
-func withStackdriverTracing(f func()) {
-	sd, err := stackdriver.NewExporter(stackdriver.Options{
-		Location: func() string {
-			s, ok := os.LookupEnv("STACKDRIVER_LOCATION")
-			if ok {
-				return s
-			}
-			s, _ = os.Hostname()
-			return s
-		}(),
-		ProjectID: os.Getenv("GOOGLE_CLOUD_PROJECT"),
-	})
-	if err != nil {
-		log.Warningf("error creating the stackdriver exporter: %v", err)
-		goto noExporter
-	}
-	// It is imperative to invoke flush before your main function exits
-	defer sd.Flush()
-
-	if s, ok := os.LookupEnv("OCTRACE_SAMPLE_PROB"); ok {
-		f, err := strconv.ParseFloat(s, 64)
-		if err != nil {
-			log.Errorf("parsing OC trace sample probability: %v", err)
-		} else {
-			trace.ApplyConfig(trace.Config{
-				DefaultSampler: trace.ProbabilitySampler(f),
-			})
-		}
-	}
-
-	// Register it as a trace exporter
-	trace.RegisterExporter(sd)
-noExporter:
-	f()
 }
 
 func mainRet() int {
